test(table): move validation bypass tests to search.test.tsx and cleanup
